Generar token

+++ src/user/user.controller.ts
@@ -1,13 +1,23 @@
-import { Body, Controller, Delete, Get, Param, Post, Put } from '@nestjs/common';
+import { Body, Controller, Delete, Get, Param, Post, Put, UseGuards } from '@nestjs/common';
+import { ApiTags } from '@nestjs/swagger';
+import { InjectRolesBuilder, RolesBuilder } from 'nest-access-control';
+import { AppResource, AppRoles } from 'src/app.roles';
+import { Auth } from 'src/common/helpers/decorators/auth.decorator';
+import { User } from 'src/common/helpers/decorators/user.decorator';
 import { CreateUserDto } from './dtos/create-user.dto';
 import { EditUserDto } from './dtos/edit-user.dto';
+import { UserRegistrationDto } from './dtos/user-registration.dto';
+import { User as UserEntity } from './entities/user.entity';
 import { UserService } from './user.service';
 
+@ApiTags('Users routes')
 @Controller('user')
 export class UserController {
 
     constructor(
-        private readonly userService: UserService
+        private readonly userService: UserService,
+        @InjectRolesBuilder()
+        private readonly rolesBuilder: RolesBuilder,
     ) {}
 
     @Get()
@@ -24,6 +34,20 @@ export class UserController {
         return { data }
     }
 
+    @Post('register')
+    async publicRegistration(@Body() dto: UserRegistrationDto) {
+        const data = await this.userService.createOne({
+            ...dto,
+            roles: [AppRoles.AUTHOR],
+          });
+          return { message: 'User registered', data };
+    }
+
+    @Auth({
+        possession: 'any',
+        action: 'create',
+        resource: AppResource.USER
+    })
     @Post()
     async createOne(
         @Body() dto: CreateUserDto //Almacena el contenido del cuerpo en el dto
@@ -32,21 +56,56 @@ export class UserController {
         return { message: 'User created', data }
     }
 
+    @Auth({
+        possession: 'own',
+        action: 'update',
+        resource: AppResource.USER
+    })
     @Put(':id')
     async editOne(
         @Param('id') id: number,
-        @Body() dto: EditUserDto
+        @Body() dto: EditUserDto,
+        @User() user: UserEntity
     ) {
-        const data = await this.userService.editOne(id, dto)
+        let data
+
+        if(this.rolesBuilder
+            .can(user.roles)
+            .updateAny(AppResource.USER)
+            .granted
+            ) {
+                console.log('Admin')
+                //ADMIN
+                data = await this.userService.editOne(id, dto);
+            } else {
+                console.log('Author')
+                //AUTHOR
+                const { roles, ...rest } = dto;
+                data = await this.userService.editOne(id, rest, user);
+            }
+
+        
         return { message: 'User edited', data }
     }
 
+    @Auth()
     @Delete(':id')
     async deleteOne(
-        @Param('id') id: number
+        @Param('id') id: number,
+        @User() user: UserEntity
     ) {
-        const data = await this.userService.deleteOne(id)
-        return { message: 'User deleted', data }
+        let data
+        if (this.rolesBuilder
+            .can(user.roles)
+            .updateAny(AppResource.USER)
+            .granted) {
+            // esto es un admin
+            data = await this.userService.deleteOne(id);
+          } else {
+            // esto es un author
+            data = await this.userService.deleteOne(id, user);
+          }
+          return { message: 'User deleted', data };
     }
 
 }
